requirements.txt
django
djangorestframework
python-dotenv
mysqlclient
django-cors-headers